inline: move to plugins package
