+++ klpbuild/main.py
@@ -6,11 +6,11 @@
 import sys
 
 from klpbuild.extractor import Extractor
-from klpbuild.inline import Inliner
 from klpbuild.klplib.cmd import create_parser
 from klpbuild.klplib.codestream import Codestream
 from klpbuild.klplib.ibs import IBS
 from klpbuild.klplib.ksrc import GitHelper
+from klpbuild.plugins.inline import Inliner
 from klpbuild.plugins.setup import Setup
