+++ klpbuild/plugins/inline.py
